backend/requirements.txt
# Hypercore E-commerce Backend Requirements
# Python 3.9+ required

# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1

# Database
#psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Authentication
authlib==1.3.0
PyJWT==2.8.0

# HTTP Requests (for Paystack API)
requests==2.31.0

# Email
resend==0.8.0

# Image Upload
cloudinary==1.38.0

# Environment Variables
python-dotenv==1.0.0

# Security
bcrypt==4.1.2
bleach==6.1.0

# Validation
email-validator==2.1.0

# Production Server
gunicorn==21.2.0
